+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Changes:
 
 - Added `BridgeMessageId` & `SpecVersion` types for bridges
+- Added `WrapperOpaque` encoding/decoding (as per latest Substrate)
 - Bump static metadata for latest Substrate, Polkadot & Kusama